[IMP] Add ReadonlyViewMixin
