+++ base_crapo_workflow/mixins/crapo_automata_mixins.py
@@ -12,14 +12,21 @@ from odoo import SUPERUSER_ID
 from odoo.tools.safe_eval import safe_eval
 from odoo.osv import expression
 
+from odoo.addons.base_crapo_workflow.mixins.crapo_readonly_view_mixin import (
+    ReadonlyViewMixin,
+)
 
-class ObjectWithStateMixin(object):
+
+class ObjectWithStateMixin(ReadonlyViewMixin):
     """
         Mixin class that can be used to define an Odoo Model eligible
         to be managed by a Crapo Automaton
 
         Should be use as a mixin class in existing objects
     """
+
+    _readonly_domain = "[('crapo_readonly_fields', 'like', ',{},'.format(field_name))]"
+    _readonly_fields_to_add = ["crapo_readonly_fields"]
 
     automaton = fields.Many2one(
         comodel_name="crapo.automaton",
@@ -136,56 +143,6 @@ class ObjectWithStateMixin(object):
             search_domain, order=order, access_rights_uid=SUPERUSER_ID
         )
         return states.browse(state_ids)
-
-    def _fields_view_get(
-        self, view_id=None, view_type="form", toolbar=False, submenu=False
-    ):
-        """
-            Override to add crapo_readonly_fields to arch and attrs readonly
-            on fields that could be editable
-        """
-        result = super(ObjectWithStateMixin, self)._fields_view_get(
-            view_id, view_type, toolbar, submenu
-        )
-        readonly_fields = self.fields_get(attributes=["readonly"])
-        node = etree.fromstring(result["arch"])
-        node.append(E.field(name="crapo_readonly_fields", invisible="1"))
-        self.process_field(node, readonly_fields)
-        result["arch"] = etree.tostring(node)
-        return result
-
-    def process_field(self, node, readonly_fields):
-        """
-            Add readnoly attrs if needed
-        """
-        if node.tag == "field":
-            field_name = node.get("name")
-
-            attrs = safe_eval(node.get("attrs", "{}"))
-            readonly = attrs.get("readonly") or node.get("readonly")
-            if isinstance(readonly, str):
-                readonly = safe_eval(node.get("readonly", "{}"))
-
-            # Deal with none domain value, if field is explicitly in readonly we skip
-            if not isinstance(readonly, (list, tuple)) and readonly:
-                return
-            # If there is no domain define and fields is already in readonly
-            # we skip too
-            elif readonly is None and readonly_fields[field_name]["readonly"]:
-                return
-
-            crapo_readonly = [
-                ("crapo_readonly_fields", "like", ",{},".format(field_name))
-            ]
-            if readonly:
-                crapo_readonly = expression.OR([readonly, crapo_readonly])
-
-            attrs["readonly"] = crapo_readonly
-            node.set("attrs", str(attrs))
-
-        else:
-            for child_node in node:
-                self.process_field(child_node, readonly_fields)
 
     # =================
     # Write / Create
